[ODPR-1549] - update specs
